refactor: reorganize UI package as proper design system
- Restructure UI package with component-based folder organization
- Move Button component to src/components/button/ folder
- Create proper design system documentation and README
- Update package.json with design system metadata and keywords
- Fix build script to properly copy CSS files to dist structure
- Ensure all apps continue to work with new organization
- Establish foundation for scalable component library
- Add comprehensive roadmap for future components
