--- packages/ui/src/button.module.css
@@ -1,152 +1,0 @@
-.button {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  white-space: nowrap;
-  border-radius: 0.375rem;
-  font-size: 0.875rem;
-  font-weight: 500;
-  transition: all 0.2s;
-  outline: none;
-  cursor: pointer;
-  border: none;
-  font-family: inherit;
-}
-
-.button:disabled {
-  pointer-events: none;
-  opacity: 0.5;
-}
-
-.button svg {
-  pointer-events: none;
-  flex-shrink: 0;
-}
-
-.button svg:not([class*='size-']) {
-  width: 1rem;
-  height: 1rem;
-}
-
-.button:focus-visible {
-  border: 2px solid var(--ring);
-  box-shadow: 0 0 0 3px rgba(var(--ring), 0.5);
-}
-
-.button[aria-invalid='true'] {
-  border-color: var(--destructive);
-  box-shadow: 0 0 0 3px rgba(var(--destructive), 0.2);
-}
-
-/* Variants */
-.button--default {
-  background-color: var(--primary);
-  color: var(--primary-foreground);
-}
-
-.button--default:hover {
-  background-color: rgba(var(--primary), 0.9);
-}
-
-.button--destructive {
-  background-color: var(--destructive);
-  color: white;
-}
-
-.button--destructive:hover {
-  background-color: rgba(var(--destructive), 0.9);
-}
-
-.button--destructive:focus-visible {
-  box-shadow: 0 0 0 3px rgba(var(--destructive), 0.2);
-}
-
-.button--outline {
-  border: 1px solid var(--border);
-  background-color: var(--background);
-  color: var(--foreground);
-}
-
-.button--outline:hover {
-  background-color: var(--accent);
-  color: var(--accent-foreground);
-}
-
-.button--dashed {
-  border: 2px dashed rgba(var(--muted-foreground), 0.25);
-  background-color: var(--background);
-  color: var(--muted-foreground);
-}
-
-.button--dashed:hover {
-  background-color: var(--muted);
-  color: var(--foreground);
-  border-color: rgba(var(--muted-foreground), 0.5);
-}
-
-.button--secondary {
-  background-color: var(--secondary);
-  color: var(--secondary-foreground);
-}
-
-.button--secondary:hover {
-  background-color: rgba(var(--secondary), 0.8);
-}
-
-.button--ghost {
-  background-color: transparent;
-}
-
-.button--ghost:hover {
-  background-color: var(--accent);
-  color: var(--accent-foreground);
-}
-
-.button--link {
-  background-color: transparent;
-  color: var(--primary);
-  text-decoration: underline;
-  text-underline-offset: 4px;
-}
-
-.button--link:hover {
-  text-decoration: underline;
-}
-
-/* Sizes */
-.button--default-size {
-  height: 2.25rem;
-  padding: 0.5rem 1rem;
-}
-
-.button--default-size:has(> svg) {
-  padding-left: 0.75rem;
-}
-
-.button--sm {
-  height: 2rem;
-  border-radius: 0.375rem;
-  gap: 0.375rem;
-  padding: 0 0.75rem;
-}
-
-.button--sm:has(> svg) {
-  padding-left: 0.625rem;
-}
-
-.button--lg {
-  height: 2.5rem;
-  border-radius: 0.375rem;
-  padding: 0 1.5rem;
-}
-
-.button--lg:has(> svg) {
-  padding-left: 1rem;
-}
-
-.button--icon {
-  width: 2.25rem;
-  height: 2.25rem;
-  border-radius: 0.375rem;
-}
