+++ packages/ui/src/components/button/index.ts
@@ -1,0 +1,2 @@
+export { Button } from './button';
+export type { ButtonProps } from './button';
