--- packages/ui/src/types.d.ts
@@ -1,1 +1,0 @@
-// Type declarations for the UI package
